requirements.txt
pandas
pyyaml
trafilatura
requests
beautifulsoup4
duckdb
lxml
python-dotenv
streamlit
pdfplumber
rapidfuzz
certifi
openpyxl
